+++ src/main/java/dev/cyan/travel/service/HotelService.java
@@ -3,9 +3,12 @@ package dev.cyan.travel.service;
 import dev.cyan.travel.DTO.HotelDTO;
 import dev.cyan.travel.entity.Country;
 import dev.cyan.travel.entity.Hotel;
+import dev.cyan.travel.entity.Room;
+import dev.cyan.travel.exception.CannotDeleteException;
 import dev.cyan.travel.mapper.HotelMapper;
 import dev.cyan.travel.repository.CountryRepository;
 import dev.cyan.travel.repository.HotelRepository;
+import dev.cyan.travel.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +20,7 @@ import java.util.Optional;
 public class HotelService {
     private final HotelRepository hotelRepository;
     private final CountryRepository countryRepository;
+    private final RoomRepository roomRepository;
     private final HotelMapper hotelMapper;
 
     public List<HotelDTO> getAll() {
@@ -50,8 +54,16 @@ public class HotelService {
         return hotelMapper.toDTO(modifiedHotel);
     }
 
-    public void delete(String id) {
-        hotelRepository.deleteById(id);
+    public void delete(String id) throws CannotDeleteException {
+        Hotel hotel = hotelRepository
+                .findById(id)
+                .orElseThrow();
+        List<Room> rooms = roomRepository.findRoomsByHotel(hotel);
+        if (rooms.isEmpty()) {
+            hotelRepository.deleteById(id);
+        } else {
+            throw new CannotDeleteException("Cannot delete this hotel because there are rooms in this hotel");
+        }
     }
 
     public List<HotelDTO> getHotelsByCountryId(String id) {
